Refactor DefaultAlluxioWorker get() and main()
The get() factory functionality is moved to the AlluxioWorker
interface.
The main() method is moved to a new AlluxioWorkerLauncher class.
