+++ core/server/src/main/java/alluxio/worker/package-info.java
@@ -11,9 +11,9 @@
 /**
  * Worker process and utils for working with the worker remotely.
  *
- * Main entry point for the worker is {@link alluxio.worker.AlluxioWorker#main(String[])} which gets
- * started by the alluxio start scripts. The {@link alluxio.worker.AlluxioWorker} class spins up the
- * different RPC services (thrift, data) which are mostly wrappers around
+ * Main entry point for the worker is {@link alluxio.worker.AlluxioWorkerLauncher#main(String[])}
+ * which gets started by the alluxio start scripts. The {@link alluxio.worker.AlluxioWorker}
+ * spins up the different RPC services (thrift, data) which are mostly wrappers around
  * {@link alluxio.worker.block.BlockWorker}.
  *
  * <h1>Services</h1>
